+++ app/src/main/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryEffect.kt
@@ -21,7 +21,11 @@ data object ShowOngoingHypertensionTreatmentError : NewMedicalHistoryViewEffect(
 
 data object ShowDiagnosisRequiredError : NewMedicalHistoryViewEffect()
 
+data object ShowDiagnosisOrReferralRequiredError : NewMedicalHistoryViewEffect()
+
 data object ShowHypertensionDiagnosisRequiredError : NewMedicalHistoryViewEffect()
+
+data object ShowHypertensionDiagnosisOrReferralRequiredError : NewMedicalHistoryViewEffect()
 
 data object ShowChangeDiagnosisErrorDialog : NewMedicalHistoryViewEffect()