+++ app/src/main/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryScreen.kt
@@ -18,7 +18,10 @@ import org.simple.clinic.R
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.di.injector
 import org.simple.clinic.feature.Features
-import org.simple.clinic.medicalhistory.SelectDiagnosisErrorDialog
+import org.simple.clinic.medicalhistory.SelectDiagnosisOrReferralRequiredErrorDialog
+import org.simple.clinic.medicalhistory.SelectDiagnosisRequiredErrorDialog
+import org.simple.clinic.medicalhistory.SelectHypertensionDiagnosisOrReferralRequiredErrorDialog
+import org.simple.clinic.medicalhistory.SelectHypertensionDiagnosisRequiredErrorDialog
 import org.simple.clinic.medicalhistory.SelectOngoingDiabetesTreatmentErrorDialog
 import org.simple.clinic.medicalhistory.SelectOngoingHypertensionTreatmentErrorDialog
 import org.simple.clinic.medicalhistory.ui.NewMedicalHistoryUi
@@ -110,15 +113,19 @@ class NewMedicalHistoryScreen : Fragment(), NewMedicalHistoryUiActions, HandlesB
   }
 
   override fun showDiagnosisRequiredErrorDialog() {
-    SelectDiagnosisErrorDialog.show(activity.supportFragmentManager, true)
+    SelectDiagnosisRequiredErrorDialog.show(activity.supportFragmentManager)
+  }
+
+  override fun showDiagnosisOrReferralRequiredErrorDialog() {
+    SelectDiagnosisOrReferralRequiredErrorDialog.show(activity.supportFragmentManager)
   }
 
   override fun showHypertensionDiagnosisRequiredErrorDialog() {
-    MaterialAlertDialogBuilder(requireContext())
-        .setTitle(getString(R.string.select_diagnosis_error_diagnosis_required))
-        .setMessage(getString(R.string.select_diagnosis_error_enter_diagnosis_hypertension))
-        .setPositiveButton(getString(R.string.select_diagnosis_error_ok), null)
-        .show()
+    SelectHypertensionDiagnosisRequiredErrorDialog.show(activity.supportFragmentManager)
+  }
+
+  override fun showHypertensionDiagnosisRequiredOrReferralErrorDialog() {
+    SelectHypertensionDiagnosisOrReferralRequiredErrorDialog.show(activity.supportFragmentManager)
   }
 
   override fun showChangeDiagnosisErrorDialog() {