+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryUpdateTest.kt
@@ -223,7 +223,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false,
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -232,7 +233,7 @@ class PatientSummaryUpdateTest {
   }
 
   @Test
-  fun `when there are patient summary changes and at least one measurement is present and no diagnosis is recorded and diabetes management is enabled, then clicking on back must show diagnosis error`() {
+  fun `when no diagnosis is recorded and diabetes management and screening is enabled, then clicking on back must show diagnosis error`() {
     val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
 
     updateSpec
@@ -250,16 +251,45 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = true
         ))
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(ShowDiagnosisError(true) as PatientSummaryEffect)
+            hasEffects(ShowDiagnosisOrReferralRequiredError as PatientSummaryEffect)
         ))
   }
 
   @Test
-  fun `when there are patient summary changes and at least one measurement is present and no diagnosis is recorded and diabetes management is disabled, then show diagnosis error`() {
+  fun `when no diagnosis is recorded and diabetes management is enabled and screening is disabled, then clicking on back must show diagnosis error`() {
+    val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForBackClickLoaded(
+            hasPatientMeasurementDataChangedSinceScreenCreated = true,
+            hasAppointmentChangeSinceScreenCreated = false,
+            countOfRecordedBloodPressures = 1,
+            countOfRecordedBloodSugars = 0,
+            medicalHistory = TestData.medicalHistory(
+                uuid = UUID.fromString("1d686e68-b64b-40ae-ba33-614c81853389"),
+                patientUuid = patientUuid,
+                diagnosedWithHypertension = Unanswered,
+                hasDiabetes = Unanswered
+            ),
+            canShowPatientReassignmentWarning = false,
+            prescribedDrugs = emptyList(),
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowDiagnosisRequiredError as PatientSummaryEffect)
+        ))
+  }
+
+  @Test
+  fun `when no diagnosis is recorded and diabetes management is disabled and screening is enabled, then show diagnosis error`() {
     val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementDisabled)
 
     updateSpec
@@ -277,11 +307,40 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = true
         ))
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(ShowDiagnosisError(false) as PatientSummaryEffect)
+            hasEffects(ShowHypertensionDiagnosisOrReferralRequiredError)
+        ))
+  }
+
+  @Test
+  fun `when no diagnosis is recorded and diabetes management and screening is disabled, then show diagnosis error`() {
+    val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementDisabled)
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForBackClickLoaded(
+            hasPatientMeasurementDataChangedSinceScreenCreated = true,
+            hasAppointmentChangeSinceScreenCreated = false,
+            countOfRecordedBloodPressures = 1,
+            countOfRecordedBloodSugars = 0,
+            medicalHistory = TestData.medicalHistory(
+                uuid = UUID.fromString("9da22d50-9a5e-4c78-b451-39cc8535fec9"),
+                patientUuid = patientUuid,
+                diagnosedWithHypertension = Unanswered,
+                hasDiabetes = Unanswered
+            ),
+            canShowPatientReassignmentWarning = false,
+            prescribedDrugs = emptyList(),
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowHypertensionDiagnosisRequiredError)
         ))
   }
 
@@ -305,7 +364,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -333,7 +393,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -361,7 +422,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -388,7 +450,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -416,7 +479,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -443,7 +507,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -452,7 +517,7 @@ class PatientSummaryUpdateTest {
   }
 
   @Test
-  fun `when at least one measurement is present and diagnosis is not recorded and diabetes management is enabled, clicking on save must show diagnosis error`() {
+  fun `when diagnosis is not recorded and diabetes management and screening is enabled, clicking on save must show diagnosis or referral error`() {
     val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
 
     updateSpec
@@ -470,16 +535,45 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = true
         ))
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(ShowDiagnosisError(true) as PatientSummaryEffect)
+            hasEffects(ShowDiagnosisOrReferralRequiredError as PatientSummaryEffect)
         ))
   }
 
   @Test
-  fun `when at least one measurement is present and diagnosis is not recorded and diabetes management is disabled, clicking on save must show diagnosis error`() {
+  fun `when diagnosis is not recorded and diabetes management is enabled and screening is disabled, clicking on save must show diagnosis error`() {
+    val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForDoneClickLoaded(
+            hasPatientMeasurementDataChangedSinceScreenCreated = true,
+            hasAppointmentChangeSinceScreenCreated = false,
+            countOfRecordedBloodPressures = 1,
+            countOfRecordedBloodSugars = 0,
+            medicalHistory = TestData.medicalHistory(
+                uuid = UUID.fromString("13c1b8b9-7109-4b2f-ad01-3532f9be6766"),
+                patientUuid = patientUuid,
+                diagnosedWithHypertension = Unanswered,
+                hasDiabetes = Unanswered
+            ),
+            canShowPatientReassignmentWarning = false,
+            prescribedDrugs = emptyList(),
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowDiagnosisRequiredError as PatientSummaryEffect)
+        ))
+  }
+
+  @Test
+  fun `when diagnosis is not recorded and diabetes management is disabled and screening is enabled, clicking on save must show diagnosis error`() {
     val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementDisabled)
 
     updateSpec
@@ -497,11 +591,40 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = true
         ))
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(ShowDiagnosisError(false) as PatientSummaryEffect)
+            hasEffects(ShowHypertensionDiagnosisOrReferralRequiredError)
+        ))
+  }
+
+  @Test
+  fun `when diagnosis is not recorded and diabetes management and screening is disabled, clicking on save must show diagnosis error`() {
+    val model = defaultModel.currentFacilityLoaded(facilityWithDiabetesManagementDisabled)
+
+    updateSpec
+        .given(model)
+        .whenEvent(DataForDoneClickLoaded(
+            hasPatientMeasurementDataChangedSinceScreenCreated = true,
+            hasAppointmentChangeSinceScreenCreated = false,
+            countOfRecordedBloodPressures = 1,
+            countOfRecordedBloodSugars = 0,
+            medicalHistory = TestData.medicalHistory(
+                uuid = UUID.fromString("e6514821-3fb4-4b78-b20a-e8ab6f856c0b"),
+                patientUuid = patientUuid,
+                diagnosedWithHypertension = Unanswered,
+                hasDiabetes = Unanswered
+            ),
+            canShowPatientReassignmentWarning = false,
+            prescribedDrugs = emptyList(),
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
+        ))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowHypertensionDiagnosisRequiredError)
         ))
   }
 
@@ -525,7 +648,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -916,7 +1040,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -943,7 +1068,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -970,7 +1096,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -997,7 +1124,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1028,7 +1156,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -1055,7 +1184,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -1082,7 +1212,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownMeasurementsWarningDialog()),
@@ -1109,7 +1240,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1345,7 +1477,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1372,7 +1505,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1404,7 +1538,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1436,7 +1571,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1473,7 +1609,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1670,7 +1807,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = true,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1701,7 +1839,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = true,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -1973,7 +2112,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = listOf("amlodipine"),
                 diabetesPrescriptions = emptyList()
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2008,7 +2148,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = emptyList(),
                 diabetesPrescriptions = listOf("metformin")
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2047,7 +2188,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = listOf("amlodipine"),
                 diabetesPrescriptions = listOf("metformin")
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2082,7 +2224,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = listOf("amlodipine"),
                 diabetesPrescriptions = emptyList()
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2117,7 +2260,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = emptyList(),
                 diabetesPrescriptions = listOf("metformin")
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2156,7 +2300,8 @@ class PatientSummaryUpdateTest {
             diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions(
                 htnPrescriptions = listOf("amlodipine"),
                 diabetesPrescriptions = listOf("metformin")
-            )
+            ),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasModel(model.shownDiagnosisWarningDialog()),
@@ -2840,7 +2985,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),
@@ -2867,7 +3013,8 @@ class PatientSummaryUpdateTest {
             ),
             canShowPatientReassignmentWarning = false,
             prescribedDrugs = emptyList(),
-            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty()
+            diagnosisWarningPrescriptions = DiagnosisWarningPrescriptions.empty(),
+            isScreeningEnabled = false
         ))
         .then(assertThatNext(
             hasNoModel(),